added Transactions.
Now you can send an array, with multiple SQL commands, as a transaction.

A transaction is a single unit of work. If a transaction is successful, all data modifications made during the transaction are committed and become a permanent part of the database. If a transaction encounters errors and needs to be canceled or reversed, all data modifications are erased.